The Config File Update

+++ mcutils/update.sh
@@ -1,18 +1,18 @@
 #!/usr/bin/env bash
 
-# THIS IS THE UPDATE SCRIPT!
-# This is a utility function (ends in .sh), it should not be called directly by the command line user.
-# This should not need to be updated outside of updates to functionality of the program itself.
+### [UPDATE WORKER] ################################################
+#   Description:  Worker script that performs the update tasks.
+#   Parameters:   1: (required) Server directory name
+#   Global Conf:  GAMEVER
 
-# PARAMS:
-# None
-
-# SETTINGS:
-#
-# Game version: Version of Minecraft to check for updates for.
-GAMEVER=1.19.3
+############################ [WARNING] ############################
+##    No part of this script is designed to be user-editable.    ##
+##  This script is OVERWRITTEN any time a C.U.M. update is run.  ##
+###################################################################
 
 
+
+# PERMISSIONS GUARD
 
 if [[ "$EUID" != 0 ]]; then
     echo "STARTUP WORKER: Insufficient privilege. Skipping startup."
@@ -21,25 +21,34 @@ fi
 
 
 
+# SCRIPT VARIABLES
 
+GAMEVER=$(cat /minecraft/cum.conf | grep "^[^#;]" | grep 'GAMEVER=' | cut -d '=' -f 2)
+
+
+
+####
 
 
 
 # CLEAR TMP DIRECTORY
-rm -rf /minecraft/tmp
+
+if [ -d "/minecraft/tmp" ]; then rm -rf /minecraft/tmp; fi
 mkdir /minecraft/tmp
 
 
 
-# GET AND PROCESS AVAILABLE LOAD INFORMATION
+# GET AND PROCESS JSON FOR LATEST STABLE PAPER BUILD
+
 echo "(worker) UPDATE: Getting latest build information for $GAMEVER..."
 
-curl -s -X 'GET' "https://api.papermc.io/v2/projects/paper/versions/$GAMEVER/builds" -H 'accept: application/json' -o /minecraft/tmp/builds.json > /dev/null
+curl -s -X 'GET' "https://api.papermc.io/v2/projects/paper/versions/$GAMEVER/builds" -H 'accept: application/json' -o /minecraft/tmp/builds.json
 jq '.builds[-1] | {build, channel, downloads}' /minecraft/tmp/builds.json > /minecraft/tmp/latest.json
 
 
 
 # DETERMINE THE LATEST BUILD NUMBER
+
 FOUND=false
 I=1
 
@@ -59,6 +68,7 @@ done
 
 
 # STORE INFORMATION ABOUT THE LATEST STABLE VERSION
+
 BUILD=$(jq '.build' /minecraft/tmp/latest.json)
 NAME=$(jq '.downloads.application.name' /minecraft/tmp/latest.json | tail -c +2 | head -c -2)
 CHECKSUM=$(jq '.downloads.application.sha256' /minecraft/tmp/latest.json | tail -c +2 | head -c -2)
@@ -66,23 +76,27 @@ CHECKSUM=$(jq '.downloads.application.sha256' /minecraft/tmp/latest.json | tail 
 
 
 # DOWNLOAD THE LATEST JAR FILE
+
 echo "(worker) UPDATE: Downloading latest stable jar file..."
 
 wget -q https://api.papermc.io/v2/projects/paper/versions/$GAMEVER/builds/$BUILD/downloads/$NAME -P /minecraft/tmp/
 
 
 
-# TEST THE CHECKSUM
+# TEST THE CHECKSUM (GUARD)
+
 TESTSUM="$(sha256sum /minecraft/tmp/$NAME | cut -d " " -f 1)"
 
 if [ $TESTSUM != $CHECKSUM ]; then
         echo "UPDATE: Checksum comparison of download failed. This is a security risk. Update failed."
+        rm -rf /minecraft/tmp
         exit
 fi
 
 
 
 # MOVE THE JAR TO THE /minecraft/jars/ FOLDER AND MAKE IT EXECUTABLE
+
 echo "(worker) UPDATE: Setting jar file as new global default."
 
 mv /minecraft/tmp/$NAME /minecraft/jars/
@@ -91,15 +105,15 @@ chmod +x /minecraft/jars/$NAME
 
 
 # UPDATE THE LATEST FILE FOR GLOBAL SERVER UPDATES
+
 echo "/minecraft/jars/$NAME" > /minecraft/jars/latest
 
 
 
 # CLEAN UP
+
 rm -rf /minecraft/tmp
-mkdir /minecraft/tmp
 
 
 
-# WE'RE DONE!
 echo "(worker) UPDATE: Update complete. Changes won't take effect until a server restart."
